+++ src/aeon3odt/odt_character_sheets.py
@@ -16,29 +16,24 @@ class OdtCharacterSheets(OdtFile):
     DESCRIPTION = 'Character sheets'
     SUFFIX = '_character_sheets'
 
-    fileHeader = OdtFile.CONTENT_XML_HEADER + '''<text:p text:style-name="Title">$Title</text:p>
-<text:p text:style-name="Subtitle">$AuthorName</text:p>
-'''
+    fileHeader = OdtFile.CONTENT_XML_HEADER
 
     characterTemplate = '''<text:h text:style-name="Heading_20_2" text:outline-level="2">$Title$FullName$AKA</text:h>
-<text:section text:style-name="Sect1" text:name="CrID:$ID">
-<text:h text:style-name="Heading_20_3" text:outline-level="3">Description</text:h>
-<text:section text:style-name="Sect1" text:name="CrID_desc:$ID">
-<text:p text:style-name="Text_20_body">$Desc</text:p>
-</text:section>
-<text:h text:style-name="Heading_20_3" text:outline-level="3">Bio</text:h>
-<text:section text:style-name="Sect1" text:name="CrID_bio:$ID">
+
+<text:p text:style-name="Text_20_body"><text:span text:style-name="Emphasis">$Tags</text:span></text:p>
+
+<text:p text:style-name="Text_20_body" />
 <text:p text:style-name="Text_20_body">$Bio</text:p>
-</text:section>
-<text:h text:style-name="Heading_20_3" text:outline-level="3">Goals</text:h>
-<text:section text:style-name="Sect1" text:name="CrID_goals:$ID">
+
+<text:p text:style-name="Text_20_body" />
 <text:p text:style-name="Text_20_body">$Goals</text:p>
-</text:section>
-<text:h text:style-name="Heading_20_3" text:outline-level="3">Notes</text:h>
-<text:section text:style-name="Sect1" text:name="CrID_notes:$ID">
-<text:p text:style-name="Text_20_body">$Notes</text:p>
-</text:section>
-</text:section>
+
+<text:p text:style-name="Text_20_body" />
+<text:p text:style-name="Text_20_body">$Desc</text:p>
+
+<text:p text:style-name="Text_20_body" />
+<text:p text:style-name="Text_20_body"><text:span text:style-name="Emphasis">$Notes</text:span></text:p>
+
 '''
 
     fileFooter = OdtFile.CONTENT_XML_FOOTER